`isNonEmptyTextNode` returns `true` for text node contained `&nbsp;` symbol.
We are using that api to determine that parent node of passed node is content editable node. If user removes all content from `contenteditable` node we inject `&nbsp;` symbol to that node in order to help editor to determine that node is `contenteditable` after page reload. That why we want to return `true` when node contains `&nbsp` symbol.